feat: move icons into icon folder
