+++ src/components/ui/icons/clipboard-icon.tsx
